Fixed Examples for tests
