+++ examples/example_003.py
@@ -14,3 +14,4 @@ def draw_pattern():
     for x, y in grid(spacing=10):
         if random() > 0.5:
             rectangle(x, y, 8, 8, "#F34060")
+        print(x, y)
