+++ examples/example_005.py
@@ -1,0 +1,6 @@
+from pypen import *
+
+
+def start():
+    for x, y in pixels():
+        rectangle(x, y, 1, 1, (random(255), random(255), random(255)))
