+++ pypen/utils/math.py
@@ -12,6 +12,7 @@ from math import e as E
 
 from random import random as py_random
 
+
 def random(x=1):
     return py_random() * x
 
@@ -32,5 +33,6 @@ def lerp(start, end, t):
 def mix(a, b, t):
     return lerp(a, b, t)
 
+
 def remap(n, start1, stop1, start2, stop2):
-  return ((n-start1)/(stop1-start1))*(stop2-start2)+start2
+    return ((n-start1)/(stop1-start1))*(stop2-start2)+start2
